requirements.txt
streamlit>=1.28.0
easyocr>=1.7.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
pydantic>=2.0.0
